разделяет логику создания миниатюр от добавления миниатюр в блок
